+++ ThirdPartyAdapters/applovin/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## AppLovin Android Mediation Adapter Changelog
 
+#### Next version
+- Removed old interstitial API implementation. Note: New interstitial API
+  implementation already exists in the adapter and GMA SDK will use that.
+
 #### Version 13.4.0.0
 - Verified compatibility with AppLovin SDK 13.4.0.